requirements.txt
pysrt
google-cloud-texttospeech
google-auth
